requirements.txt
fastapi[standard]==0.113.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
alembic==1.13.1
asyncpg==0.28.0
python-dotenv==1.0.0

